refactor(InputView): 공백을 입력받는 예외도 객체를 생성할 때 처리하도록 변경
입력받는 InputView는 어떤 입력이든 입력을 받는 하나의 책임을 가져야 한다고 판단
객체를 생성할 때 예외처리
