+++ src/main/java/domain/RaceWinner.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class RaceWinner {
-    List<Car> raceWinners = new ArrayList<>();
+    private List<Car> raceWinners = new ArrayList<>();
 
     public void addRaceWinner(Car car) {
         raceWinners.add(car);
